Move to separate test class
